Moved FullScan to a separate class.
Changes:
* Refactored code and moved the whole full scan to a separate class
(will separate all parts of the app by classes in next commits);
* Improved overall app speed (Less CPU/RAM usage);
* Added <summary> to some methods;
* Simplified scanning, so they all use same methods (it means you can
create a new type of scan or embed it into your own scanner in a few
minutes, documentation will be available closer to release).

Known problems:
* Loading speed is awful with slow / no internet connection (persists
for a long time);
* Logs aren't working correctly with files containing "(", ")", "|",
"{", "}", "[", "]" in their names (persists for a long time).

Possible issues:
* Some parts may have the wrong locale;
* Background Worker may still remain working even after canceling it;
* UI thread may freeze if you enable OK logging.